Moved HCP menu to it's own template file
By moving the HCP menu to it's own template file, we can just include
that template on each HCP page. It'd have been nicer to be able to
modify the layout.html all the pages currently already use, but I didn't
know how to get the current role via Thymeleaf. This fixes #93